util/utilbot/pyproject.toml
[tool.poetry]
name = "gd-utilities"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = "^0.17.1"
discord = "^1.0.1"
uvloop = "^0.15.2"
aerich = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
